requirements.txt
appdirs==1.4.4
asgiref==3.7.2
baron==0.10.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==4.2.7
django-allauth==0.58.2
django-cache-url==3.4.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-environs==0.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
environs==9.5.0
gunicorn==21.2.0
idna==3.4
marshmallow==3.20.1
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
redbaron==0.9.2
requests==2.31.0
requests-oauthlib==1.3.1
rply==0.7.8
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
